[Fix] refactor
